drop webdav http methods
