+++ lib/route.js
@@ -549,8 +549,8 @@ function normalizeAndValidateMethod (method) {
     throw new FST_ERR_ROUTE_METHOD_INVALID()
   }
   method = method.toUpperCase()
-  if (!this[kAcceptedHTTPMethods].bodyless.has(method)
-    && !this[kAcceptedHTTPMethods].bodywith.has(method)) {
+  if (!this[kAcceptedHTTPMethods].bodyless.has(method) &&
+    !this[kAcceptedHTTPMethods].bodywith.has(method)) {
     throw new FST_ERR_ROUTE_METHOD_NOT_SUPPORTED(method)
   }
