+++ test/http-methods/copy.test.js
@@ -3,13 +3,14 @@
 const t = require('tap')
 const test = t.test
 const sget = require('simple-get').concat
-const fastify = require('..')()
+const fastify = require('../../fastify')()
+fastify.acceptHTTPMethod('COPY')
 
-test('can be created - unlock', t => {
+test('can be created - copy', t => {
   t.plan(1)
   try {
     fastify.route({
-      method: 'UNLOCK',
+      method: 'COPY',
       url: '*',
       handler: function (req, reply) {
         reply.code(204).send()
@@ -25,13 +26,13 @@ fastify.listen({ port: 0 }, err => {
   t.error(err)
   t.teardown(() => { fastify.close() })
 
-  test('request - unlock', t => {
+  test('request - copy', t => {
     t.plan(2)
     sget({
-      url: `http://localhost:${fastify.server.address().port}/test/a.txt`,
-      method: 'UNLOCK',
+      url: `http://localhost:${fastify.server.address().port}/test.txt`,
+      method: 'COPY',
       headers: {
-        'Lock-Token': 'urn:uuid:a515cfa4-5da4-22e1-f5b5-00a0451e6bf7'
+        Destination: '/test2.txt'
       }
     }, (err, response, body) => {
       t.error(err)
