+++ test/http-methods/lock.test.js
@@ -3,7 +3,8 @@
 const t = require('tap')
 const test = t.test
 const sget = require('simple-get').concat
-const fastify = require('..')()
+const fastify = require('../../fastify')()
+fastify.acceptHTTPMethod('LOCK', { hasBody: true })
 
 const bodySample = `<?xml version="1.0" encoding="utf-8" ?>
         <D:lockinfo xmlns:D='DAV:'>
