+++ test/http-methods/move.test.js
@@ -3,7 +3,8 @@
 const t = require('tap')
 const test = t.test
 const sget = require('simple-get').concat
-const fastify = require('..')()
+const fastify = require('../../')()
+fastify.acceptHTTPMethod('MOVE')
 
 test('shorthand - move', t => {
   t.plan(1)
