+++ test/http-methods/proppatch.test.js
@@ -3,7 +3,8 @@
 const t = require('tap')
 const test = t.test
 const sget = require('simple-get').concat
-const fastify = require('..')()
+const fastify = require('../../')()
+fastify.acceptHTTPMethod('PROPPATCH', { hasBody: true })
 
 const bodySample = `<?xml version="1.0" encoding="utf-8" ?>
         <D:propertyupdate xmlns:D="DAV:"
