+++ test/http-methods/search.test.js
@@ -3,7 +3,8 @@
 const t = require('tap')
 const sget = require('simple-get').concat
 const test = t.test
-const fastify = require('..')()
+const fastify = require('../../fastify')()
+fastify.acceptHTTPMethod('SEARCH', { hasBody: true })
 
 const schema = {
   response: {
