+++ test/http-methods/trace.test.js
@@ -2,7 +2,8 @@
 
 const t = require('tap')
 const test = t.test
-const fastify = require('..')()
+const fastify = require('../../fastify')()
+fastify.acceptHTTPMethod('TRACE')
 
 test('shorthand - trace', t => {
   t.plan(1)
