+++ test/http-methods/unlock.test.js
@@ -3,16 +3,17 @@
 const t = require('tap')
 const test = t.test
 const sget = require('simple-get').concat
-const fastify = require('..')()
+const fastify = require('../../fastify')()
+fastify.acceptHTTPMethod('UNLOCK')
 
-test('can be created - mkcol', t => {
+test('can be created - unlock', t => {
   t.plan(1)
   try {
     fastify.route({
-      method: 'MKCOL',
+      method: 'UNLOCK',
       url: '*',
       handler: function (req, reply) {
-        reply.code(201).send()
+        reply.code(204).send()
       }
     })
     t.pass()
@@ -25,14 +26,17 @@ fastify.listen({ port: 0 }, err => {
   t.error(err)
   t.teardown(() => { fastify.close() })
 
-  test('request - mkcol', t => {
+  test('request - unlock', t => {
     t.plan(2)
     sget({
-      url: `http://localhost:${fastify.server.address().port}/test/`,
-      method: 'MKCOL'
+      url: `http://localhost:${fastify.server.address().port}/test/a.txt`,
+      method: 'UNLOCK',
+      headers: {
+        'Lock-Token': 'urn:uuid:a515cfa4-5da4-22e1-f5b5-00a0451e6bf7'
+      }
     }, (err, response, body) => {
       t.error(err)
-      t.equal(response.statusCode, 201)
+      t.equal(response.statusCode, 204)
     })
   })
 })
